Integration of state machine in GameManagerCore
 Integrate the state machine in the current GameManager/GameManagerCore pattern
 Update documentation
