+++ Fightship Arena/Assets/Scripts/Managers/GameManagement/IGameManagerCore.cs
@@ -1,7 +1,12 @@
-﻿namespace FightShipArena.Assets.Scripts.Managers.GameManagement
+﻿using UnityEngine.InputSystem;
+
+namespace FightShipArena.Assets.Scripts.Managers.GameManagement
 {
     public interface IGameManagerCore
     {
-        
+        void OnAwake();
+        void OnStart();
+
+        void OnPauseResumeGame(InputAction.CallbackContext context);
     }
 }
