+++ Fightship Arena/Assets/Scripts/Managers/SceneManagement/SceneManager.cs
@@ -10,13 +10,9 @@ namespace FightShipArena.Assets.Scripts.Managers.SceneManagement
 {
     public class SceneManager : MyMonoBehaviour
     {
-        //private PlayerInput _playerInput;
-
         void Start()
         {
             Debug.Log("Scene started");
-
-           //_playerInput = GetComponent<PlayerInput>();
 
         }
     }
